Shift from per-class files to one file for related classes. Add a proper unit test for SnsHandler.
