+++ test/test.py
@@ -1,4 +1,6 @@
 # -- coding: utf-8 --
+from __future__ import absolute_import
+
 from unittest import main, TestCase
 
 from tempfile import mkdtemp
